+++ autoload.php
@@ -15,11 +15,18 @@
  * 
  */
 
+// Define __DIR__ for PHP version < 5.3 (should be already included in config file)
+if (! defined('__DIR__')) define('__DIR__', dirname(__FILE__));
+
+// Load Dolibarr environment
+if (false === (@include_once __DIR__ . '/../main.inc.php')) {  // From htdocs directory (dolibase is in htdocs directory)
+	if (false === (@include_once __DIR__ . '/../../main.inc.php')) { // From module directory (dolibase is in module directory)
+		require_once __DIR__ . '/../../../main.inc.php'; // From "custom" directory (module itself is in custom directory)
+	}
+}
+
 // Dolibarr detection
 if (! defined('DOL_VERSION')) die('Dolibase::autoload::error Dolibarr detection failed.');
-
-// Define __DIR__ for PHP version < 5.3
-if (! defined('__DIR__')) define('__DIR__', dirname(__FILE__));
 
 // Load Dolibase constants
 require_once __DIR__ . '/const.php';