+++ test/books/config.php
@@ -10,7 +10,7 @@ $dolibase_config['module_name']     = 'Books';
 
 $dolibase_config['module_desc']     = 'Books Managment Module';
 
-$dolibase_config['module_version']  = '1.0.0'; // Possible values: 'development', 'experimental', 'dolibarr' or version
+$dolibase_config['module_version']  = '1.1.0'; // Possible values: 'development', 'experimental', 'dolibarr' or version
 
 $dolibase_config['module_number']   = '450000'; // e.: 550000 (avoid small numbers because they are used for core modules)
 
@@ -78,6 +78,12 @@ $dolibase_config['num_model_prefix'] = 'BK';
  * Load Dolibase
  */
 
-if (! defined('DOLIBASE_VERSION') && false === (@include_once DOL_DOCUMENT_ROOT.'/dolibase/autoload.php')) { // From htdocs directory
-	dol_include_once('/'.$dolibase_config['module_folder'].'/dolibase/autoload.php'); // From module directory
+// Define __DIR__ for PHP version < 5.3
+if (! defined('__DIR__')) define('__DIR__', dirname(__FILE__));
+
+// Load Dolibase only if not already loaded by another module
+if (! defined('DOLIBASE_VERSION') && false === (@include_once __DIR__ . '/../dolibase/autoload.php')) { // From htdocs directory
+	if (false === (@include_once __DIR__ . '/../../dolibase/autoload.php')) { // From "custom" directory
+		require_once __DIR__ . '/dolibase/autoload.php'; // From module directory
+	}
 }